# solution/docs/api.yaml
openapi: 3.1.0
info:
  title: Advertisement Service
  description: PROD Advertisement Serviice API
  version: v1.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /time/advance:
    post:
      tags:
        - Utils
      summary: Установка текущей даты
      description: Устанавливает текущий день в системе в заданную дату.
      operationId: setCurrentDate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCurrentDate"
        required: true
      responses:
        "200":
          description: Текущая дата обновлена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SetCurrentDate"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SetCurrentDate"
  /moderation:
    post:
      tags:
        - Utils
      summary: Управление модерацией
      description: Включает/выключает модерацию.
      operationId: setModerationState
      parameters:
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: Модерация включена/выключена
          content:
            '*/*':
              schema:
                type: boolean
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: boolean
  /moderation/banned_word:
    post:
      tags:
        - Utils
      summary: Добавление слова в список запрещенных
      description: Добавляет слово в список запрещенных.
      operationId: addBannedWord
      parameters:
        - name: word
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Слово добавлено в список запрещенных
          content:
            '*/*':
              schema:
                type: boolean
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: boolean
    delete:
      tags:
        - Utils
      summary: Удаление слова из списка запрещенных
      description: Удаляет слово из списка запрещенных.
      operationId: removeBannedWord
      parameters:
        - name: word
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Слово удалено из списка запрещенных
          content:
            '*/*':
              schema:
                type: boolean
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: boolean
...
  /clients/bulk:
    post:
      tags:
        - Clients
      summary: Массовое создание/обновление клиентов
      description: Создаёт новых или обновляет существующих клиентов
      operationId: createOrUpdateClient
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ClientRequest"
        required: true
      responses:
        "201":
          description: Успешное создание/обновление клиентов
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientRequest"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientRequest"
  /advertisers/{advertiserId}/campaigns:
    get:
      tags:
        - Campaigns
      summary: Получение рекламных кампаний рекламодателя c пагинацией
      description: Возвращает список рекламных кампаний для указанного рекламодателя
        с пагинацией.
      operationId: getAllByAdvertiserId
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: Список рекламных кампаний рекламодателя.
          content:
            '*/*':
              schema:
                type: object
        "204":
          description: Список рекламных кампаний рекламодателя пуст.
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Рекламодатель не найден.
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
        - Campaigns
      summary: Создание рекламной кампании
      description: Создаёт новую рекламную кампанию для указанного рекламодателя.
      operationId: create_1
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementCreateRequest"
        required: true
      responses:
        "201":
          description: Рекламная кампания успешно создана.
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: object
...
  /stats/campaigns/{advertisementId}:
    get:
      tags:
        - Statistics
      summary: Получение статистики по рекламной кампании
      description: "\n\t\tВозвращает агрегированную статистику (показы, переходы,\
        \ затраты и конверсию) для заданной рекламной кампании.\n\t\t"
      operationId: advertisementStatistic
      parameters:
        - name: advertisementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Статистика по рекламной кампании успешно получена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Statistic"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Statistic"
  /stats/campaigns/{advertisementId}/daily:
    get:
      tags:
        - Statistics
      summary: Получение ежедневной статистики по рекламной кампании
      description: Возвращает массив ежедневной статистики для указанной рекламной
        кампании.
      operationId: advertisementDailyStatistic
      parameters:
        - name: advertisementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ежедневная статистика по рекламной кампании успешно получена.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyStatistic"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyStatistic"
  /stats/advertisers/{advertiserId}/campaigns:
    get:
      tags:
        - Statistics
      summary: Получение агрегированной статистики по всем кампаниям рекламодателя
      description: "\n\t\t\tВозвращает сводную статистику по всем рекламным кампан\
        иям, принадлежащим заданному рекламодателю.\n\t\t"
      operationId: advertiserStatistic
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Агрегированная статистика по всем кампаниям рекламодателя успешно
            получена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Statistic"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Statistic"
  /stats/advertisers/{advertiserId}/campaigns/daily:
    get:
      tags:
        - Statistics
      summary: Получение ежедневной агрегированной статистики по всем кампаниям рекламодателя
      description: "\n\t\t\tВозвращает массив ежедневной сводной статистики по все\
        м рекламным кампаниям заданного рекламодателя.\n\t\t"
      operationId: advertiserDailyStatistic
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ежедневная агрегированная статистика успешно получена.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyStatistic"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyStatistic"
  /clients/{id}:
    get:
      tags:
        - Clients
      summary: Получение клиента по ID
      description: Возвращает информацию о клиенте по его ID.
      operationId: getClientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Информация о клиенте успешно получена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClientRequest"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClientRequest"
        "404":
          description: Клиент не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClientRequest"
...
  /advertisers/{advertiserId}/campaigns/{advertisementId}/image:
    get:
      tags:
        - Campaigns
      summary: Получение картинки рекламной кампании
      description: Возвращает картинку рекламной кампании в формате JPEG.
      operationId: getImage
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: advertisementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Картинка успешно получена.
          content:
            image/jpeg: { }
        "204":
          description: Рекламная кампания не имеет картинки.
          content:
            '*/*':
              schema:
                type: string
                format: byte
        "404":
          description: Рекламная кампания не найдена.
          content:
            '*/*':
              schema:
                type: string
                format: byte
...
components:
  schemas:
...
    Targeting:
      type: object
      properties:
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - ALL
        age_from:
          type: integer
          format: int32
          minimum: 0
        age_to:
          type: integer
          format: int32
          minimum: 0
        location:
          type: string
...
    SetCurrentDate:
      type: object
      properties:
        current_date:
          type: integer
          format: int32
          minimum: 0
      required:
        - current_date
...
    ClientRequest:
      type: object
      properties:
        client_id:
          type: string
          format: uuid
        login:
          type: string
        age:
          type: integer
          format: int32
          maximum: 130
          minimum: 0
        location:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
      required:
        - age
        - client_id
        - gender
        - location
        - login
    AdvertisementCreateRequest:
      type: object
      properties:
        impressions_limit:
          type: integer
          format: int32
          minimum: 1
        clicks_limit:
          type: integer
          format: int32
          minimum: 0
        cost_per_impression:
          type: number
          format: double
          minimum: 0
        cost_per_click:
          type: number
          format: double
          minimum: 0
        ad_title:
          type: string
        ad_text:
          type: string
        start_date:
          type: integer
          format: int32
          minimum: 0
        end_date:
          type: integer
          format: int32
          minimum: 0
        targeting:
          $ref: "#/components/schemas/Targeting"
        image_url:
          type: string
      required:
        - ad_text
        - ad_title
        - clicks_limit
        - cost_per_click
        - cost_per_impression
        - end_date
        - impressions_limit
        - start_date
        - targeting
...
    Statistic:
      type: object
      properties:
        impressions_count:
          type: integer
          format: int32
        clicks_count:
          type: integer
          format: int32
        conversion:
          type: number
          format: double
        spent_impressions:
          type: number
          format: double
        spent_clicks:
          type: number
          format: double
        spent_total:
          type: number
          format: double
      required:
        - clicks_count
        - conversion
        - impressions_count
        - spent_clicks
        - spent_impressions
        - spent_total
    DailyStatistic:
      type: object
      properties:
        impressions_count:
          type: integer
          format: int32
        clicks_count:
          type: integer
          format: int32
        conversion:
          type: number
          format: double
        spent_impressions:
          type: number
          format: double
        spent_clicks:
          type: number
          format: double
        spent_total:
          type: number
          format: double
        date:
          type: integer
          format: int32
      required:
        - clicks_count
        - conversion
        - date
        - impressions_count
        - spent_clicks
        - spent_impressions
        - spent_total
...