# solution/src/main/kotlin/com/intezya/solution/dto/Statistic.kt
package com.intezya.solution.dto

import com.fasterxml.jackson.annotation.JsonProperty

data class Statistic(
	@field:JsonProperty("impressions_count")
	val impressionsCount: Int,
	@field:JsonProperty("clicks_count")
	val clicksCount: Int,
	@field:JsonProperty("conversion")
	val conversion: Double,
	@field:JsonProperty("spent_impressions")
	val spentImpressions: Double,
	@field:JsonProperty("spent_clicks")
	val spentClicks: Double,
	@field:JsonProperty("spent_total")
	val spentTotal: Double,
)